# services/core/src/swagger/swagger.yaml
openapi: "3.0.3"
info:
  title: GlobalTransact API
  version: "2.0.0"
  description: |
    International money transfer platform API.

    All endpoints require Bearer JWT authentication unless marked as public.

    **Roles**: client | agent | admin | compliance_officer | super_admin
servers:
  - url: /api/v1
    description: Core API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /transfers:
    post:
      tags: [Transfers]
      summary: Initiate a new money transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipientName, recipientMsisdn, recipientCountry, sendCurrency, recvCurrency, sendAmount]
              properties:
                recipientName: { type: string }
                recipientMsisdn: { type: string }
                recipientCountry: { type: string }
                sendCurrency: { type: string, enum: [RWF, USD, EUR, GBP, KES, UGX, GHS] }
                recvCurrency: { type: string, enum: [RWF, USD, EUR, GBP, KES, UGX, GHS] }
                sendAmount: { type: number, minimum: 1 }
                fundingMethod: { type: string, enum: [momo, cash, wallet, bank], default: momo }
                payoutMethod: { type: string, enum: [momo, cash, bank], default: momo }
                beneficiaryId: { type: string, format: uuid }
      responses:
        201: { description: Transfer initiated }
        422: { description: Blocked by fraud engine or insufficient funds }
        503: { description: FX rates stale or no liquidity }

    get:
      tags: [Transfers]
      summary: List own transfers
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        200: { description: Transfer list }
  /transfers/{urc}:
    get:
      tags: [Transfers]
      summary: Get transfer by URC
      parameters:
        - in: path
          name: urc
          required: true
          schema: { type: string }
      responses:
        200: { description: Transfer details }
        404: { description: Not found }
  /transfers/{id}/approve:
    post:
      tags: [Transfers]
      summary: Admin approve transfer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Transfer approved }
  /transfers/{id}/reject:
    post:
      tags: [Transfers]
      summary: Admin reject transfer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Transfer rejected }
...